requirements.txt
fastapi
uvicorn
mysql-connector-python
sentence-transformers
numpy
scikit-learn
requests
python-dotenv